+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+### Removed
+- [BREAKING] dropped support for a broken 'dominant' border colour
+
 ## [0.99.4] 22.11.2024
 ### Added
 - Better test coverage for straighten operation